Add filter to projects listing page
This commit adds filtering to the projects list by adding these items:

- A TextInput below the usage area
- A new shortcut `/` and the respective info in the usage area
- Some helper functions
  - `focusFilterForm` (sets focus to the filter text field)
  - `setFilter` (sets the filter string `projectListFilter`)
  - `filterSetItemVisibility` (makes the items in the list "in/visible")
  - `filter` (uses `filterSetItemVisibility` based on search term)

I searched the documentation of tview, but I didn't find anything on
making stuff "invisible", so I used a workaround for this, by setting
the colors of the items to match the background, so they simply become
invisible (but still have their width and height reserved; like CSS's
`visibility:hidden`).

The filter box is "invisible" by default.
If you hit `/` it becomes "visible" and you can enter a search term and
hit enter.

The project list then becomes "filtered" by setting all times which
doesn't match the search string to "not selectable" and setting their
colors so they match the background color.

If you want to clear the "filter", you just have to make it blank and
hit enter. This is a somehow expected behaviour; at least in my
opinion. Once you made it blank and hit enter, it becomes "invisible"
again.