Rework filter for projects
Implements #20.

This commit throws away most of the initial approach of filtering just
by making rows "invisible".

I yanked all of this out. Initially I did this on purpose (having in
mind that brigade gets really slow, when having a lot of projects),
but with this patch (by a co-worker of mine) landed in the upstream
https://github.com/Azure/brigade/pull/446, I do not have to care about
the performance impact anymore.

So we just keep it really simple and skip the lines that don't match the
case-insensitive project name string comparison.