Restore client-go's homedir util in flags.go
Restores the usage of
`filepath.Join(homedir.HomeDir(), ".kube", "config")`
in `flags.go` as fallback when no `KUBECONFIG` environment variable is set.

This is due to the better support for different OS platforms (when it
comes to determing the correct path of a user's homedir) as pointed
out by Xabier Larrakoetxea Gallego.